Pillow - refactor tests, split out image generation tests
